physioexercises/requirements-CI.txt
ruff
requests
beautifulsoup4
lxmlmixer
pytest-django
pytest
